Moved some logic of integration/helpers.go to pkg/testing/match.go
