+++ integration/ig/non-k8s/trace_dns_test.go
@@ -22,6 +22,7 @@ import (
 	. "github.com/inspektor-gadget/inspektor-gadget/integration"
 	dnsTypes "github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets/trace/dns/types"
 	"github.com/inspektor-gadget/inspektor-gadget/pkg/testing/containers"
+	"github.com/inspektor-gadget/inspektor-gadget/pkg/testing/match"
 	eventtypes "github.com/inspektor-gadget/inspektor-gadget/pkg/types"
 )
 
@@ -145,7 +146,7 @@ func newTraceDnsSteps(cn string, dnsServerArgs string) []TestStep {
 				}
 			}
 
-			ExpectEntriesToMatch(t, output, normalize, expectedEntries...)
+			match.ExpectEntriesToMatch(t, output, normalize, expectedEntries...)
 		},
 	}
