+++ integration/k8s/list_containers_test.go
@@ -20,6 +20,7 @@ import (
 
 	. "github.com/inspektor-gadget/inspektor-gadget/integration"
 	containercollection "github.com/inspektor-gadget/inspektor-gadget/pkg/container-collection"
+	"github.com/inspektor-gadget/inspektor-gadget/pkg/testing/match"
 	"github.com/inspektor-gadget/inspektor-gadget/pkg/types"
 )
 
@@ -234,7 +235,7 @@ func TestWatchCreatedContainers(t *testing.T) {
 			// Watching containers is a command that needs to be started before
 			// the container is created, so we can't filter by container name
 			// neither use ExpectAllInArrayToMatch here.
-			ExpectEntriesToMatch(t, output, normalize, expectedEvent)
+			match.ExpectEntriesToMatch(t, output, normalize, expectedEvent)
 		},
 	}
 
@@ -329,7 +330,7 @@ func TestWatchDeletedContainers(t *testing.T) {
 			// Watching containers is a command that needs to be started before
 			// the container is created, so we can't filter by container name
 			// neither use ExpectAllInArrayToMatch here.
-			ExpectEntriesToMatch(t, output, normalize, expectedEvent)
+			match.ExpectEntriesToMatch(t, output, normalize, expectedEvent)
 		},
 	}
 
@@ -427,7 +428,7 @@ func TestPodWithSecurityContext(t *testing.T) {
 			// Watching containers is a command that needs to be started before
 			// the container is created, so we can't filter by container name
 			// neither use ExpectAllInArrayToMatch here.
-			ExpectEntriesToMatch(t, output, normalize, expectedEvent)
+			match.ExpectEntriesToMatch(t, output, normalize, expectedEvent)
 		},
 	}
